* master: (627 commits)
  Handling unsub'd feed with active IFTTT trigger.
  Adding dark story header border color for android. For #546.
  Don't record read stories for aggregated mark reads.
  Adding unsub'd feeds to read stories view.
  Fixing Subscribe button for tryfeeds.
  Fixing IFTTT saved story tags.
  Moving classifiers when recreating destroyed feeds.
  Code to recreateaccidentally destroyed feeds.
  Adding SSL certificate assembly fab command.
  Skip saved feeds which have fake unread counts.
  Fix style of share bar text for dark theme.
  Apply dark theme to classifier dialog.
  Fixing saved story view for users with preference to open first unread story, thereby switching them out of saved mode and into unread mode.
  Update ShareDialogFragment to use AlertDialog.Builder and default theme
  Adding copy_db_settings to fabfile.
  Apply dark theme to replies.
  Update ReplyDialogFragment to use AlertDialog.Builder and default themes.
  Fixing specified image sizes.
  Only premiums users can share multiple stories from a single feed per day.
  Faster delete user (for combating spam).
  ...

+++ clients/ios/Other Sources/InAppSettingsKit/Models/IASKSettingsStore.h
@@ -24,7 +24,7 @@
 - (void)setBool:(BOOL)value      forKey:(NSString*)key;
 - (void)setFloat:(float)value    forKey:(NSString*)key;
 - (void)setDouble:(double)value  forKey:(NSString*)key;
-- (void)setInteger:(int)value    forKey:(NSString*)key;
+- (void)setInteger:(NSInteger)value    forKey:(NSString*)key;
 - (void)setObject:(id)value      forKey:(NSString*)key;
 - (BOOL)boolForKey:(NSString*)key;
 - (float)floatForKey:(NSString*)key;
